# digital-laundry-backend/src/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": { "title": "Laundry Management API", "version": "1.0.0" },
  "servers": [{ "url": "/" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/admin/orders": {
      "get": {
        "summary": "List orders",
        "parameters": [
          {
            "name": "hostelCode",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "uniqueLaundryId",
            "in": "query",
            "schema": { "type": "string" }
          },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          {
            "name": "dateFrom",
            "in": "query",
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "dateTo",
            "in": "query",
            "schema": { "type": "string", "format": "date" }
          },
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          {
            "name": "pageSize",
            "in": "query",
            "schema": { "type": "integer" }
          },
          { "name": "sort", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Paged" } }
      }
    },
    "/api/orders": {
      "post": {
        "summary": "Create order",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["orderDate", "items"],
                "properties": {
                  "orderDate": { "type": "string" },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": ["itemType", "quantity"],
                      "properties": {
                        "itemType": { "type": "string" },
                        "quantity": { "type": "integer" },
                        "remark": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Order created" },
          "400": { "description": "Validation" }
        }
      },
      "get": {
        "summary": "List orders",
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          {
            "name": "pageSize",
            "in": "query",
            "schema": { "type": "integer" }
          },
          { "name": "sort", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Paged" } }
      }
    },
    "/api/orders/{id}": {
      "get": {
        "summary": "Get order",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Order with items/images/qr/complaints" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api/orders/{id}/complete": {
      "post": {
        "summary": "Mark completed (PENDING → COMPLETED)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Completed" },
          "400": { "description": "Invalid transition" }
        }
      }
    },
    "/api/orders/{orderId}/items/{itemId}/images": {
      "post": {
        "summary": "Upload order item images (JSON array of base64/URLs)",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["images"],
                "properties": {
                  "images": { "type": "array", "items": { "type": "string" } }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Uploaded" } }
      }
    },
    "/api/orders/{id}/qrcode": {
      "get": {
        "summary": "Get QR payload",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "QR payload" } }
      }
    },
...
    "/api/history/orders/{id}/history": {
      "get": {
        "summary": "Order history",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "Status entries" } }
      }
    },
    "/api/history/complaints/{id}/history": {
      "get": {
        "summary": "Complaint history",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "Status entries" } }
      }
    },
    "/api/orders/{id}/history": {
      "get": {
        "summary": "Order history (alias)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "Status entries" } }
      }
    },
    "/api/complaints/{id}/history": {
      "get": {
        "summary": "Complaint history (alias)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "Status entries" } }
      }
    }
  }
}